+++ src/Client.php
@@ -123,23 +123,39 @@ final class Client implements ClientInterface
     /**
      * {@inheritdoc}
      */
-    public function captureMessage(string $message, ?Severity $level = null, ?Scope $scope = null): ?EventId
+    public function captureMessage(string $message, ?Severity $level = null, ?Scope $scope = null/*, ?EventHint $hint = null*/): ?EventId
     {
+        $hint = \func_num_args() > 3 ? func_get_arg(3) : null;
+
+        if (null !== $hint && !$hint instanceof EventHint) {
+            throw new \InvalidArgumentException(sprintf('The $hint argument must be an instance of the "%s" class. Got: "%s".', EventHint::class, get_debug_type($hint)));
+        }
+
         $event = Event::createEvent();
         $event->setMessage($message);
         $event->setLevel($level);
 
-        return $this->captureEvent($event, null, $scope);
+        return $this->captureEvent($event, $hint, $scope);
     }
 
     /**
      * {@inheritdoc}
      */
-    public function captureException(\Throwable $exception, ?Scope $scope = null): ?EventId
+    public function captureException(\Throwable $exception, ?Scope $scope = null/*, ?EventHint $hint = null*/): ?EventId
     {
-        return $this->captureEvent(Event::createEvent(), EventHint::fromArray([
-            'exception' => $exception,
-        ]), $scope);
+        $hint = \func_num_args() > 2 ? func_get_arg(2) : null;
+
+        if (null !== $hint && !$hint instanceof EventHint) {
+            throw new \InvalidArgumentException(sprintf('The $hint argument must be an instance of the "%s" class. Got: "%s".', EventHint::class, get_debug_type($hint)));
+        }
+
+        $hint = $hint ?? new EventHint();
+
+        if (null === $hint->exception) {
+            $hint->exception = $exception;
+        }
+
+        return $this->captureEvent(Event::createEvent(), $hint, $scope);
     }
 
     /**
@@ -170,8 +186,9 @@ final class Client implements ClientInterface
     /**
      * {@inheritdoc}
      */
-    public function captureLastError(?Scope $scope = null): ?EventId
+    public function captureLastError(?Scope $scope = null/*, ?EventHint $hint = null*/): ?EventId
     {
+        $hint = \func_num_args() > 1 ? func_get_arg(1) : null;
         $error = error_get_last();
 
         if (null === $error || !isset($error['message'][0])) {
@@ -180,7 +197,7 @@ final class Client implements ClientInterface
 
         $exception = new \ErrorException(@$error['message'], 0, @$error['type'], @$error['file'], @$error['line']);
 
-        return $this->captureException($exception, $scope);
+        return $this->captureException($exception, $scope, $hint);
     }
 
     /**
@@ -267,7 +284,7 @@ final class Client implements ClientInterface
 
         if (!$isTransaction) {
             $previousEvent = $event;
-            $event = ($this->options->getBeforeSendCallback())($event);
+            $event = ($this->options->getBeforeSendCallback())($event, $hint);
 
             if (null === $event) {
                 $this->logger->info('The event will be discarded because the "before_send" callback returned "null".', ['event' => $previousEvent]);