+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Deprecate the `logger` option (#1167)
+- Pass the event hint from the `capture*()` methods down to the `before_send` callback (#1138)
 
 ## 3.1.2 (2021-01-08)